ToDoMate/mainApp/serializers.py
from rest_framework import serializers
from .models import Schedule


class ScheduleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Schedule
        fields = ('id','title','content','schdule_date','remind_date','isActive')
        
        