ToDoMate/requirements.txt
Package             Version
------------------- -------
asgiref             3.6.0
Django              4.1.7
django-cors-headers 3.14.0
djangorestframework 3.14.0
Pillow              9.4.0
pip                 22.3.1
pytz                2023.3
setuptools          65.5.0
sqlparse            0.4.3
tzdata              2022.7
